+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ New features
 Bugfixes
 ~~~~~~~~
 
+* Fix REST API server response if the HTTP request fails.
 * Fix updating Markdown native content in text format.
 * Fix ``--format`` option with the export operation.
 * Fix parsing of text formatted solution.